Extract Item Recipe Card into a separate component
