+++ src/App.js
@@ -29,6 +29,7 @@ import axios from "axios";
 // import getCharacteristicImage from "./utility/characteristicsMap";
 import AutocompleteSearchBar from "./components/AutocompleteSearchBar";
 import ItemStatsCard from "./components/ItemStatsCard";
+import ItemRecipeCard from "./components/ItemRecipeCard";
 
 // API URL
 const Dofus3AllEquipment = `https://api.dofusdu.de/dofus3/v1/en/items/equipment/all`;
@@ -151,97 +152,7 @@ function App() {
 
         <ItemStatsCard item={selectedEquipmentData} />
 
-        {/* Recipes Card */}
-        <Box
-          sx={{
-            // minWidth: 275,
-            width: "100%",
-            marginTop: "20px",
-            boxShadow: "0px 0px 10px 0px rgba(0, 0, 0, 0.5)",
-          }}
-        >
-          <Card
-            variant="outlined"
-            sx={{ backgroundColor: "#1a1a1a", color: "white" }}
-          >
-            <CardContent>
-              <div>
-                <Typography
-                  gutterBottom
-                  sx={{ fontSize: 14, color: "lightcyan" }}
-                >
-                  {selectedEquipmentData?.name}
-                </Typography>
-                <Typography gutterBottom sx={{ color: "red", fontSize: 12 }}>
-                  {selectedEquipmentData?.type?.name}
-                </Typography>
-                <div>
-                  <img
-                    src={selectedEquipmentData?.image_urls.sd}
-                    alt={selectedEquipmentData?.name}
-                    style={{ width: "100px", height: "100px" }}
-                  />
-                </div>
-              </div>
-              <div>
-                <div>
-                  {selectedEquipmentData?.recipe?.map((resource) => {
-                    let resourceData = data.allResources.find(
-                      (res) => res.ankama_id === resource.item_ankama_id
-                    );
-                    if (resourceData === undefined) {
-                      // try to search for the resource in the allEquipmentData array
-                      resourceData = data.allEquipment.find(
-                        (res) => res.ankama_id === resource.item_ankama_id
-                      );
-                    }
-                    // TODO: if resourceData still undefined, search for consumables as part of the recipe
-                    return resourceData === undefined ? (
-                      <Typography
-                        key={resource?.item_ankama_id}
-                        variant="body1"
-                        style={{ color: "red" }}
-                      >
-                        {"Unknown Resource"} x{" "}
-                        <span style={{ color: "lightsalmon" }}>
-                          {resource.quantity}
-                        </span>
-                      </Typography>
-                    ) : (
-                      <Stack
-                        direction="row"
-                        spacing={1}
-                        key={resource?.item_ankama_id}
-                        style={{
-                          marginBottom: "10px",
-                          display: "flex",
-                          alignItems: "center",
-                          justifyContent: "center",
-                        }}
-                      >
-                        <img
-                          src={resourceData?.image_urls.sd}
-                          alt={resourceData?.name}
-                          style={{ width: "20px", height: "20px" }}
-                        />
-                        <Typography variant="body1">
-                          {resourceData?.name} x{" "}
-                          <span style={{ color: "lightsalmon" }}>
-                            {resource.quantity}
-                          </span>
-                        </Typography>
-                      </Stack>
-                    );
-                  })}
-                </div>
-              </div>
-            </CardContent>
-            {/* <CardActions>
-              <Button size="small">Learn More</Button>
-            </CardActions> */}
-          </Card>
-        </Box>
-        {/* ==== Recipes Card ==== */}
+        <ItemRecipeCard item={selectedEquipmentData} data={data} />
       </Container>
     </div>
   );
